requirements.txt
faster-coco-eval>=1.6.5,<2
pyyaml>=6.0.2,<7
tensorboard>=2.19.0,<3
scipy>=1.15.2,<2
calflops>=0.3.2,<0.4
transformers>=4.49.0,<5
opencv-python>=4.11.0.86,<5
loguru>=0.7.3,<0.8
matplotlib
mlflow-skinny
azureml-mlfow